migate collab service to TS
