+++ services/collaboration-service/package.json
@@ -3,10 +3,18 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node app.js",
-    "swagger-autogen": "node swagger.js"
+    "swagger-autogen": "node swagger.js",
+    "start": "ts-node src/app.ts",
+    "build": "tsc",
+    "start:prod": "node dist/app.js"
   },
   "dependencies": {
+    "@types/body-parser": "^1.19.3",
+    "@types/express": "^4.17.17",
+    "@types/node": "^20.6.3",
+    "@types/socket.io": "^3.0.2",
+    "@types/swagger-ui-express": "^4.1.3",
+    "@types/uuid": "^9.0.4",
     "body-parser": "^1.20.2",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
@@ -16,6 +24,8 @@
     "socket.io": "^4.7.2",
     "swagger-autogen": "^2.23.5",
     "swagger-ui-express": "^5.0.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2",
     "uuid": "^9.0.1"
   }
 }
