+++ services/collaboration-service/src/routes/index.html
